Refactor SLSListener into modular components with separate source files
